research-agent/requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
tavily-python>=0.3.0